Refactor file names to kebab-case naming convension
