+++ sdk/package/src/models/discovery-data.ts
